+++ source/extensions/lightspeed.trex.properties_pane.logic.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.2]
+### Changed
+- Improved node type description display in properties panel
+
 ## [1.2.1]
 ### Added
 - Added StagePrimPicker integration for OmniGraph target relationship inputs